requirements.txt
absl-py==1.4.0
addict==2.4.0
albumentations==1.3.1
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
decorator==4.4.2
easydict==1.10
face-detection==0.2.2
fonttools==4.41.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.56.2
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
importlib-metadata==6.8.0
importlib-resources==6.0.0
joblib==1.3.1
kiwisolver==1.4.4
lazy_loader==0.3
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.2
moviepy==1.0.3
networkx==3.1
numpy==1.25.1
oauthlib==3.2.2
opencv-python==4.8.0.74
opencv-python-headless==4.8.0.74
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
proglog==0.1.10
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
qudida==0.0.4
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorboardX==2.6.1
threadpoolctl==3.2.0
tifffile==2023.7.18
torch==1.8.1+cu111
torchaudio==0.8.1
torchvision==0.9.1+cu111
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.3.6
zipp==3.16.2
